% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{back_propagation_training_rcpp}
\alias{back_propagation_training_rcpp}
\title{The title of -back_propagation_training_rcpp-}
\usage{
back_propagation_training_rcpp(i, h, o, learning_rate, activation_func, data)
}
\arguments{
\item{i}{Numeric scalar. Number of input nodes.}

\item{h}{Numeric scalar. Number of hidden nodes.}

\item{o}{Numeric scalar. Number of output nodes.}

\item{learning_rate}{Numeric scalar. Learning rate of the algorithm.}

\item{activation_func}{Character (either "sigmoid" or "ReLU").}

\item{data}{R data frame with X columns and Y labels.}
}
\value{
A list of weight and bias matrices.
}
\description{
Here is a brief description
}
\details{
Computes the weight and bias matrices for the nodes of the neural network with \code{i} # of input nodes, \code{h} # of hidden nodes, and \code{o} # of output nodes.
}
\examples{
set.seed(100)
data <- data.frame(X1 = 1:100, X2 = 2:101, Y = sample(c(0,1), 100, replace=TRUE))
data_as_matrix <- as.matrix(data)
nn_model_rcpp <- back_propagation_training_rcpp(i=2, h=2, o=1, learning_rate=0.01,
activation_func="sigmoid", data=data_as_matrix)

}
