% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derivative_ReLU.R
\name{derivative_ReLU}
\alias{derivative_ReLU}
\title{The title of -derivative_ReLU-}
\usage{
derivative_ReLU(a)
}
\arguments{
\item{a}{Numeric matrix.}
}
\value{
A matrix of class \code{BackPropNN_derivative_ReLU}:
}
\description{
Here is a brief description
}
\details{
Computes the derivative of ReLU function value of each element of matrix \code{a}.
}
\examples{
mat <- matrix(1:9,3,3)
derivative_ReLU(mat)

}
