% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derivative_sigmoid.R
\name{derivative_sigmoid}
\alias{derivative_sigmoid}
\title{The title of -derivative_sigmoid-}
\usage{
derivative_sigmoid(a)
}
\arguments{
\item{a}{Numeric matrix.}
}
\value{
A matrix of class \code{BackPropNN_derivative_sigmoid}:
}
\description{
Here is a brief description
}
\details{
Computes the derivative of sigmoid function value of each element of matrix \code{a}.
}
\examples{
mat <- matrix(1:9,3,3)
derivative_sigmoid(mat)

}
