% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feed_forward.R
\name{feed_forward}
\alias{feed_forward}
\title{The title of -feed_forward-}
\usage{
feed_forward(data, nn_model)
}
\arguments{
\item{data}{R data frame with X columns and Y labels.}

\item{nn_model}{A list of weight and bias matrices along with other information about fitted nn model.}
}
\value{
A vector of class \code{BackPropNN_feed_forward}:
}
\description{
Here is a brief description
}
\details{
Computes the predicted value of the X columns of \code{data}.
}
\examples{
data <- data.frame(X1 = 1:10, X2 = 2:11, Y = sample(c(0,1), 10, replace=TRUE))
nn_model <- back_propagation_training(i, h, o, learning_rate, activation_func, data)
feed_forward(data, nn_model)

}
