% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{feed_forward_rcpp}
\alias{feed_forward_rcpp}
\title{The title of -feed_forward_rcpp-}
\usage{
feed_forward_rcpp(data, nn_model)
}
\arguments{
\item{data}{R data frame with X columns and Y labels.}

\item{nn_model}{A list of weight and bias matrices along with other information about fitted nn model.}
}
\value{
A vector.
}
\description{
Here is a brief description
}
\details{
Computes the predicted value of the X columns of \code{data}.
}
\examples{
data <- data.frame(X1 = 1:10, X2 = 2:11, Y = sample(c(0,1), 10, replace=TRUE))
data_as_matrix <- as.matrix(data)
nn_model_rcpp <- back_propagation_training_rcpp(i=2, h=2, o=1, learning_rate=0.01,
activation_func="sigmoid", data=data_as_matrix)
feed_forward_rcpp(data, nn_model_rcpp)

}
