% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{map}
\alias{map}
\title{The title of -map-}
\usage{
map(func, a)
}
\arguments{
\item{func}{R function.}

\item{a}{Numeric matrix.}
}
\value{
A matrix (of same dimension as matrix a) of class \code{BackPropNN_map}:
}
\description{
Here is a brief description
}
\details{
Computes the function value (func) of each element of matrix \code{a}.
}
\examples{
mat <- matrix(1:9,3,3)
func <- function(x){
return(x+2)
}
map(func, mat)

}
