% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReLU.R
\name{ReLU}
\alias{ReLU}
\title{The title of -ReLU-}
\usage{
ReLU(a)
}
\arguments{
\item{a}{Numeric matrix.}
}
\value{
A matrix of class \code{BackPropNN_ReLU}:
}
\description{
Here is a brief description
}
\details{
Computes the ReLU function value of each element of matrix \code{a}.
}
\examples{
mat <- matrix(1:9,3,3)
ReLU(mat)

}
