% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigmoid.R
\name{sigmoid}
\alias{sigmoid}
\title{The title of -sigmoid-}
\usage{
sigmoid(a)
}
\arguments{
\item{a}{Numeric matrix.}
}
\value{
A matrix of class \code{BackPropNN_sigmoid}:
}
\description{
Here is a brief description
}
\details{
Computes the sigmoid function value of each element of matrix \code{a}.
}
\examples{
mat <- matrix(1:9,3,3)
sigmoid(mat)

}
