% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sigmoid_rcpp}
\alias{sigmoid_rcpp}
\title{The title of -sigmoid_rcpp-}
\usage{
sigmoid_rcpp(a)
}
\arguments{
\item{a}{Numeric matrix.}
}
\value{
A matrix.
}
\description{
Here is a brief description
}
\details{
Computes the sigmoid function value of each element of matrix \code{a}.
}
\examples{
mat <- matrix(1:9,3,3)
sigmoid_rcpp(mat)

}
